Rename loader files
