+++ glue_solar/core/sunpy_maps/__init__.py
@@ -1,1 +1,1 @@
-from .sunpy_maps_loader import *
+from .loader import *
