+++ glue_solar/core/sunpy_maps/loader.ui
