+++ glue_solar/instruments/iris/loader.ui
